Move package collection JSON models to a separate module
Motivation:
Some publishers may opt to generate package collections on their own, which will require the JSON models. It might not be desirable to pull in all the dependencies that come with PackageCollections module.

Modifications:
Move PackageCollections/JSONModel into its own PackageCollectionJSONModel module and make it a product.

Result:
Package collection JSON models can be used with minimal dependencies.
