+++ Sources/PackageCollectionJSONModel/JSONPackageCollection+v1.swift
@@ -1,7 +1,7 @@
 /*
  This source file is part of the Swift.org open source project
 
- Copyright (c) 2020 Apple Inc. and the Swift project authors
+ Copyright (c) 2020-2021 Apple Inc. and the Swift project authors
  Licensed under Apache License v2.0 with Runtime Library Exception
 
  See http://swift.org/LICENSE.txt for license information
@@ -10,9 +10,6 @@
 
 import struct Foundation.Date
 import struct Foundation.URL
-import TSCUtility
-
-import PackageModel
 
 extension JSONPackageCollectionModel {
     public enum V1 {}
@@ -252,115 +249,66 @@ extension JSONPackageCollectionModel.V1 {
     }
 }
 
-// MARK: - Validations
+// MARK: -  Copy `PackageModel.ProductType` to minimize the module's dependencies
 
 extension JSONPackageCollectionModel.V1 {
-    public struct Validator {
-        public let configuration: Configuration
+    /// The type of product.
+    public enum ProductType: Equatable {
+        /// The type of library.
+        public enum LibraryType: String, Codable {
+            /// Static library.
+            case `static`
 
-        public init(configuration: Configuration = .init()) {
-            self.configuration = configuration
+            /// Dynamic library.
+            case dynamic
+
+            /// The type of library is unspecified and should be decided by package manager.
+            case automatic
         }
 
-        public func validate(collection: Collection) -> [ValidationMessage]? {
-            var messages = [ValidationMessage]()
+        /// A library product.
+        case library(LibraryType)
 
-            let packages = collection.packages
-            // Stop validating if collection doesn't pass basic checks
-            if packages.isEmpty {
-                messages.append(.error("A collection must contain at least one package.", property: "packages"))
-            } else if packages.count > self.configuration.maximumPackageCount {
-                messages.append(.warning("The collection has (\(packages.count)) packages, which is more than the recommended maximum (\(self.configuration.maximumPackageCount)) and extra data might be ignored.", property: "packages"))
-            } else {
-                packages.forEach { self.validate(package: $0, messages: &messages) }
-            }
+        /// An executable product.
+        case executable
 
-            guard messages.isEmpty else {
-                return messages
-            }
+        /// A test product.
+        case test
+    }
+}
 
-            return nil
+extension JSONPackageCollectionModel.V1.ProductType: Codable {
+    private enum CodingKeys: String, CodingKey {
+        case library, executable, test
+    }
+
+    public func encode(to encoder: Encoder) throws {
+        var container = encoder.container(keyedBy: CodingKeys.self)
+        switch self {
+        case .library(let a1):
+            var unkeyedContainer = container.nestedUnkeyedContainer(forKey: .library)
+            try unkeyedContainer.encode(a1)
+        case .executable:
+            try container.encodeNil(forKey: .executable)
+        case .test:
+            try container.encodeNil(forKey: .test)
         }
+    }
 
-        // TODO: validate package url?
-        private func validate(package: Collection.Package, messages: inout [ValidationMessage]) {
-            let packageID = PackageIdentity(url: package.url.absoluteString).description
-
-            // Check for duplicate versions
-            let nonUniqueVersions = Dictionary(grouping: package.versions, by: { $0.version }).filter { $1.count > 1 }.keys
-            if !nonUniqueVersions.isEmpty {
-                messages.append(.error("Duplicate version(s) found in package \(packageID): \(nonUniqueVersions).", property: "package.versions"))
-            }
-
-            var nonSemanticVersions = [String]()
-            let semanticVersions: [TSCUtility.Version] = package.versions.compactMap {
-                let semver = TSCUtility.Version(string: $0.version)
-                if semver == nil {
-                    nonSemanticVersions.append($0.version)
-                }
-                return semver
-            }
-
-            guard nonSemanticVersions.isEmpty else {
-                messages.append(.error("Non semantic version(s) found in package \(packageID): \(nonSemanticVersions).", property: "package.versions"))
-                // The next part of validation requires sorting the semvers. Cannot continue if non-semver.
-                return
-            }
-
-            let sortedVersions = semanticVersions.sorted(by: >)
-
-            var currentMajor: Int?
-            var majorCount = 0
-            var minorCount = 0
-            for version in sortedVersions {
-                if version.major != currentMajor {
-                    currentMajor = version.major
-                    majorCount += 1
-                    minorCount = 0
-                }
-
-                guard majorCount <= self.configuration.maximumMajorVersionCount else {
-                    messages.append(.warning("Package \(packageID) includes too many major versions. Only \(self.configuration.maximumMajorVersionCount) is allowed and extra data might be ignored.", property: "package.versions"))
-                    break
-                }
-                guard minorCount < self.configuration.maximumMinorVersionCount else {
-                    // !-safe currentMajor cannot be nil at this point
-                    messages.append(.warning("Package \(packageID) includes too many minor versions for major version \(currentMajor!). Only \(self.configuration.maximumMinorVersionCount) is allowed and extra data might be ignored.", property: "package.versions"))
-                    break
-                }
-
-                minorCount += 1
-            }
-
-            package.versions.forEach { version in
-                if version.products.isEmpty {
-                    messages.append(.error("Package \(packageID) version \(version.version) does not contain any products.", property: "version.products"))
-                }
-                version.products.forEach { product in
-                    if product.targets.isEmpty {
-                        messages.append(.error("Product \(product.name) of package \(packageID) version \(version.version) does not contain any targets.", property: "product.targets"))
-                    }
-                }
-
-                if version.targets.isEmpty {
-                    messages.append(.error("Package \(packageID) version \(version.version) does not contain any targets.", property: "version.targets"))
-                }
-            }
+    public init(from decoder: Decoder) throws {
+        let values = try decoder.container(keyedBy: CodingKeys.self)
+        guard let key = values.allKeys.first(where: values.contains) else {
+            throw DecodingError.dataCorrupted(.init(codingPath: decoder.codingPath, debugDescription: "Did not find a matching key"))
         }
-
-        public struct Configuration {
-            public var maximumPackageCount: Int
-            public var maximumMajorVersionCount: Int
-            public var maximumMinorVersionCount: Int
-
-            public init(maximumPackageCount: Int? = nil,
-                        maximumMajorVersionCount: Int? = nil,
-                        maximumMinorVersionCount: Int? = nil) {
-                // TODO: where should we read defaults from?
-                self.maximumPackageCount = maximumPackageCount ?? 50
-                self.maximumMajorVersionCount = maximumMajorVersionCount ?? 2
-                self.maximumMinorVersionCount = maximumMinorVersionCount ?? 3
-            }
+        switch key {
+        case .library:
+            var unkeyedValues = try values.nestedUnkeyedContainer(forKey: key)
+            let a1 = try unkeyedValues.decode(JSONPackageCollectionModel.V1.ProductType.LibraryType.self)
+            self = .library(a1)
+        case .test:
+            self = .test
+        case .executable:
+            self = .executable
         }
     }
 }
